refactor!(core): delete the `io` feature from the core crate (#255)
* update manifests

* remove feature gates

* reorg io files
