+++ honeycomb-kernels/Cargo.toml
@@ -13,7 +13,7 @@ authors.workspace = true
 publish = true
 
 [dependencies]
-honeycomb-core = { workspace = true, features = ["io"] }
+honeycomb-core.workspace = true
 num-traits.workspace = true
 rayon = "1.10.0"
 thiserror.workspace = true
