+++ honeycomb/Cargo.toml
@@ -18,6 +18,6 @@ kernels = ["dep:honeycomb-kernels"]
 render = ["dep:honeycomb-render"]
 
 [dependencies]
-honeycomb-core = { workspace = true, features = ["io"] }
+honeycomb-core = { workspace = true }
 honeycomb-kernels = { workspace = true, optional = true }
 honeycomb-render = { workspace = true, optional = true }
